pyproject.toml
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "python-fhir-converter"
version = "0.0.19"
authors = ["Chase Stewart <chase.a.stewart@outlook.com>"]
description = "Transformation utility to translate data formats into FHIR"
readme = "README.md"
license = "MIT"
homepage = "https://github.com/chaseastewart/fhir-converter"
repository = "https://github.com/chaseastewart/fhir-converter"
documentation = "https://chaseastewart.github.io/fhir-converter/"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]
packages = [
    {include = "fhir_converter"},
    {include = "fhir_converter/py.typed"},
]

[tool.poetry.dependencies]
python = "^3.9"
python-liquid = "^1.10.2"
xmltodict = "^0.13.0"
pyjson5 = "^1.6.5"
frozendict = "^2.3.10"
importlib_resources = "^6.1.1"
psutil = "^5.9.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-mock = "^3.12.0"
pytest-cov = "^4.1.0"
types-xmltodict = "^0.13.0.3"
types-psutil = "^5.9.5.17"
flake8 = "^7.0.0"
mypy = "^1.8.0"
smokeshow = "^0.4.0"
lazydocs = "^0.4.8"
mkdocs-awesome-pages-plugin = "^2.9.2"
mkdocs-include-markdown-plugin = "^6.0.4"

[tool.poetry.scripts]
fhir_converter_cli  = "fhir_converter.__main__:entrypoint"